+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- None
+### Added
+
+- User option `Show checklist completion status` to display the completion status of all checklists in the selected note
 
 ## [1.3.0] - 2021-03-09